Disable file matching 'fuzzy' logic and ON-TOP disable
only exact directb2s filename matching to table.vpx is better than the
old method of fuzzy matching when no exact is found.  left as commented
as someone could/should make a b2s/xml setting to enabled/disable that
code.  Also bringtofront is replaced with sendtoback for backglass,
this should be a setting as well in future.